+++ Dockerfile
@@ -7,4 +7,4 @@ RUN pip install --no-cache-dir -r requirements.txt
 
 COPY server/ .
 
-CMD ["uvicorn", "api_products.main:app", "--host", "0.0.0.0", "--port", "8000"]
+CMD ["uvicorn", "api_navimall.main:app", "--host", "0.0.0.0", "--port", "8000"]
