+++ server/Tools/save_images.py
@@ -4,7 +4,7 @@ import requests
 from urllib.parse import urlparse, parse_qs
 
 # Chemins
-JSON_FILE = "../api_products/assets/json/products.json"
+JSON_FILE = "../api_navimall/assets/json/products.json"
 IMAGES_DIR = os.path.join("..", "assets", "products_images")
 
 os.makedirs(IMAGES_DIR, exist_ok=True)
