+++ server/api_navimall/path_optimization/serialization_utils.py
