+++ server/assets/layout_examples/befeac1196230ec6_metadata.json
@@ -1,0 +1,15 @@
+{
+  "layout_hash": "befeac1196230ec6",
+  "grid_shape": [
+    30,
+    40
+  ],
+  "edge_length": 100.0,
+  "statistics": {
+    "navigable": 702,
+    "obstacles": 498,
+    "pois": 0
+  },
+  "file_path": "C:/Users/romai/OneDrive - De Vinci/Bureau/Navimall/server/assets/layout_examples\\befeac1196230ec6.h5",
+  "created_with": "NaviStore Grid Editor"
+}
